requirements.txt
alembic==1.7.7
aniso8601==9.0.1
attrs==21.4.0
click==8.0.4
Flask==2.0.3
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
importlib-metadata==4.11.3
importlib-resources==5.4.0
iniconfig==1.1.1
itsdangerous==2.1.1
Jinja2==3.0.3
Mako==1.2.0
MarkupSafe==2.1.1
marshmallow==3.15.0
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pyparsing==3.0.7
pytest==7.1.1
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.32
SQLAlchemy-serializer==1.4.1
tomli==2.0.1
Werkzeug==2.0.3
zipp==3.7.0
